==> 03_Plot_Figs_2_3_S7.Rmd <==
---
title: "Script to make multipanel figures 2, 3 along with figure S7"
output:
html_notebook: default
---

```{r}
install_if <- function(x) {
  
  if(x %in% rownames(installed.packages()) == FALSE) {
  message(paste(x, "is required but not installed. Installing now"))
  Sys.sleep(1)
  install.packages(x)
  library(x)
  } else{ 
    library(x, character.only=TRUE)}
  }

install_if("ggplot2")
install_if("cowplot")
install_if("here")
install_if("ggpubr")

# make a folder to export analysis data
if(! dir.exists(here("figures"))){
  dir.create(here("figures"))
}

source(here("function_plotting_theme.R"))
```



# Figure 2

To produce this figure, you must run the code for Figure 2 a + inset in 1_Data_simlation.Rmd and the code for Figure 2 b in 2c_Experiment_analysis.nb.Rmd. The `Scenario` in Simulations must be set to 2 (weighted mean) and the sensitivity analysis (Analysis) must be toggled of. 

load figures
```{r}

load(here("figures", "Fig_2a.RData"))
load(here("figures", "Fig_2a_i.RData"))
load(here("figures", "Fig_2b.RData"))

```


arrange figures
```{r}
Fig_2a <- 
  Fig_2a +
  labs(x = "")+
  theme_meta()+
  scale_y_continuous(limits = c(-0.2,1.1))+
  labs(y = "Transgressive overyielding /\n Standardized function", x = "Habitat heterogeneity")


Fig_2a_i <- 
  Fig_2a_i +
  labs(x = "", y = "", title = "")

Fig_2a <- 
  Fig_2a +
  annotation_custom(
    ggplotGrob(Fig_2a_i),
    xmin = 0.5, xmax = 5, ymin = -0.3, ymax = 0.4
  )

Fig_2b <- 
  Fig_2b +
  labs(x = "")+
  scale_y_continuous(limits = c(-0.2,1.1))+
  labs(y = "")+
   theme_meta()

# add another figure without the BEF slope
p1 <- 
  ggpubr::ggarrange(Fig_2a + labs(x = "Habitat heterogeneity"), 
                    Fig_2b + labs(x = "Habitat heterogeneity"), 
                    ncol = 2, nrow = 1, widths = c(1, 1),
                    labels = c("A", "B"),
                    font.label = 
                      list(size = 10, 
                           color = "black", 
                           face = "bold", family = NULL))

p1

ggsave(filename = here("figures", "Fig_2.pdf"), 
       plot = p1, width = 18, height = 9, units = "cm", dpi = 450)

```

#Figure 3

To produce this figure, you must run the code for Figure 3 a,b & in Analysis.nb.Rmd. The sensitivity analysis  (Analysis) must be toggled of. 

load Figures
```{r}
load(here("figures", "Fig_3A.RData"))
load(here("figures", "Fig_3B.RData"))
```


arrange Figures
```{r}
Fig_3B <- 
  Fig_3B+
  theme_meta()+
  xlab("Strain richness") +
  theme(legend.position = "none")+
  labs(y = "")

Legend <- cowplot::get_legend(
  # create some space to the left of the legend
  Fig_3A + 
    theme(legend.position = "bottom", 
          legend.box.margin = margin(0, 0, 0, 12),
          legend.key=element_blank()) 
)

Fig_3A <- 
  Fig_3A+
  theme_meta()+
  theme(legend.position = "none")

# add the legend to the row we made earlier. Give it one-third of 
# the width of one plot (via rel_widths).

Fig3 <- plot_grid(Fig_3A, Fig_3B,
                align = "v",
                labels="AUTO",
                label_x = 0.1, 
                label_size = 10)

Fig3 <- plot_grid(Fig3, Legend, nrow = 2, rel_heights = c(2, .4))

Fig3

ggsave(plot = Fig3, filename = here("figures", "Fig_3.pdf"), 
       width = 18, height = 9, units = "cm", dpi = 450)
```


# Figure S7

note that to produce this figure you must run the code in Analysis.nb.Rmd both with and without the sensitivity analysis

load figures
```{r}
load(here("figures", "Fig_2b.RData"))

Fig_2b_with_a <- 
  Fig_2b+
  theme_meta()+
  scale_y_continuous(limits = c(0,1.1)) +
  labs(y = "Transgressive overyielding /\n Standardized function", x = "Habitat heterogeneity")

load(here("figures", "Fig_S7.RData"))

Fig_2b <- Fig_2b+
   theme_meta()+
  scale_y_continuous(limits = c(0,1.1)) +
  ylab(NULL)
```


arrange figures
```{r}
p1 <- plot_grid(Fig_2b_with_a, Fig_2b, 
                align = "v", labels="AUTO",
                label_x = 0)

p1

ggsave(plot = p1, filename = here("figures", "Fig_S7.pdf"), width = 8, height = 4)


```
